v1.3.1
git-svn-id: https://plugins.svn.wordpress.org/wp-access-areas/tags/1.3.1@992458 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/class-undisclosededitpost.php
@@ -40,9 +40,8 @@ class UndisclosedEditPost {
 	}
 	
 	static function add_post_type_columns() {
-		$can_edit_edit_cap		=  ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_edit_cap' ) ;//current_user_can( 'manage_options' );
-		$can_edit_view_cap		= ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_view_cap' ) );
-		$can_edit_comment_cap	= ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_comment_cap' ) );
+		
+		$can_edit_edit_cap		=  self::can_edit_edit_cap();
 
 		// posts
 		if ( post_type_supports( 'post' , 'comments' ) )
@@ -78,8 +77,8 @@ class UndisclosedEditPost {
 		// custom post types
 		foreach ( $post_types as $post_type ) {
 			$post_type_object = get_post_type_object( $post_type );
-			$view_col 		= $can_edit_view_cap && $post_type_object->public || $post_type_object->show_ui;
-			$comment_col	= $can_edit_comment_cap && post_type_supports( $post_type , 'comments' );
+			$view_col 		= self::can_edit_view_cap( $post_type , $post_type_object );
+			$comment_col	= self::can_edit_comment_cap( $post_type );
 
 			if ( $view_col && $comment_col ) 
 				add_filter( "manage_{$post_type}_posts_columns" , array( __CLASS__ , 'add_disclosure_columns'));
@@ -132,20 +131,21 @@ class UndisclosedEditPost {
 		global $wp_post_types;
 		if ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_view_cap' ) || current_user_can( 'wpaa_set_edit_cap' ) || current_user_can( 'wpaa_set_comment_cap' ) ) {
 			foreach ( array_keys($wp_post_types) as $post_type ) {
-				add_meta_box( 'post-disclosure' , __('Access','wpundisclosed') , array(__CLASS__,'disclosure_box_info') , $post_type , 'side' , 'high' );
-				$post_type_object 	= get_post_type_object( $post_type );
-				if ( $post_type_object->public && $post_type != 'attachment' )
-					add_meta_box( 'post-disclosure-behavior' , __('Behaviour','wpundisclosed') , array(__CLASS__,'disclosure_box_behavior') , $post_type , 'side' , 'high' );
+				add_meta_box( 'post-disclosure' , __('Access','wp-access-areas') , array(__CLASS__,'disclosure_box_info') , $post_type , 'side' , 'high' );
+				if ( self::can_edit_view_cap( $post_type ) )
+					add_meta_box( 'post-disclosure-behavior' , __('Behaviour','wp-access-areas') , array(__CLASS__,'disclosure_box_behavior') , $post_type , 'side' , 'high' );
 			}
 		}
 	}
 	// --------------------------------------------------
-	// saving post
+	// saving posts
 	// --------------------------------------------------
 	static function edit_post( $data, $postarr ) {
-
-		$post_type_object 	= get_post_type_object($data["post_type"]);
+	
+		$post_type = $data["post_type"];
+		$post_type_object 	= get_post_type_object( $post_type );
 		
+		// set default values
 		if ( ! $postarr['ID'] ) {
 			$caps = array( 
 				'post_view_cap' => 'exist',
@@ -161,40 +161,46 @@ class UndisclosedEditPost {
 			return $data;
 
 		
-		if (  ( $post_type_object->public || $post_type_object->show_ui ) && isset($postarr['post_view_cap']) && $postarr['post_view_cap'] )
-			$data['post_view_cap']	= $postarr['post_view_cap'];
+		// process user input. 
+		if ( self::can_edit_view_cap( $post_type , $post_type_object ) && isset($postarr['post_view_cap']) && $postarr['post_view_cap'] )
+			$data['post_view_cap']	= wpaa_sanitize_access_cap(  $postarr['post_view_cap'] );
 		
-		if (isset($postarr['post_edit_cap']) && $postarr['post_edit_cap']) 
-			$data['post_edit_cap']	= $postarr['post_edit_cap'];
+		if ( self::can_edit_edit_cap() && isset($postarr['post_edit_cap']) && $postarr['post_edit_cap']) 
+			$data['post_edit_cap']	= wpaa_sanitize_access_cap( $postarr['post_edit_cap'] );
 	
-		if ( post_type_supports( $data["post_type"] , 'comments' ) && isset($postarr['post_comment_cap']) && $postarr['post_comment_cap'] )
-			$data['post_comment_cap']	= $postarr['post_comment_cap'];
+		if (  self::can_edit_comment_cap( $post_type ) && isset($postarr['post_comment_cap']) && $postarr['post_comment_cap'] )
+			$data['post_comment_cap']	= wpaa_sanitize_access_cap( $postarr['post_comment_cap'] );
 		
 		return $data;
 	}
 	
+	// --------------------------------------------------
+	// saving posts, 
+	// --------------------------------------------------
 	static function set_post_behavior(  $post_ID , $post , $update ) {
-		// set page
-		if ( isset( $_POST['_wpaa_fallback_page'] ) )
-			update_post_meta( $post_ID , '_wpaa_fallback_page' , intval( $_POST['_wpaa_fallback_page'] ) );
+		// should only happen if edit_view_cap is true
+		if ( self::can_edit_view_cap( $post->post_type ) ) {
+			if ( isset( $_POST['_wpaa_fallback_page'] ) )
+				update_post_meta( $post_ID , '_wpaa_fallback_page' , intval( $_POST['_wpaa_fallback_page'] ) );
 		
-		if ( isset( $_POST['_wpaa_post_behavior'] ) ) {
-			$meta = $_POST['_wpaa_post_behavior'];
+			if ( isset( $_POST['_wpaa_post_behavior'] ) ) {
+				$meta = $_POST['_wpaa_post_behavior'];
 		
-			if ( $meta === '' ) {
-				delete_post_meta( $post_ID , '_wpaa_post_behavior' );
-			} else if ( in_array( $meta , array( '404' , 'page' , 'login' ) ) ) {
-				update_post_meta( $post_ID , '_wpaa_post_behavior' , $meta );
+				if ( $meta === '' ) {
+					delete_post_meta( $post_ID , '_wpaa_post_behavior' );
+				} else if ( in_array( $meta , array( '404' , 'page' , 'login' ) ) ) {
+					update_post_meta( $post_ID , '_wpaa_post_behavior' , $meta );
+				}
 			}
 		}
 	}
 	
 	static function edit_attachment( $attachment_ID ) {
-		$attachment = get_post($attachment_ID);
-		$post_edit_cap = isset($_POST['post_edit_cap']) ? sanitize_title($_POST['post_edit_cap']) : $attachment->post_edit_cap;
-		$post_comment_cap = isset($_POST['post_comment_cap']) ? sanitize_title($_POST['post_comment_cap']) : $attachment->post_comment_cap;
+		$attachment			= get_post($attachment_ID);
+		$post_edit_cap 		= isset($_POST['post_edit_cap']) ? wpaa_sanitize_access_cap( $_POST['post_edit_cap'] ) : $attachment->post_edit_cap;
+		$post_comment_cap	= isset($_POST['post_comment_cap']) ? wpaa_sanitize_access_cap( $_POST['post_comment_cap'] ) : $attachment->post_comment_cap;
 	
-		if ( $post_edit_cap != $attachment->post_edit_cap || $post_comment_cap != $attachment->post_comment_cap ) {
+		if ( $attachment && $post_edit_cap != $attachment->post_edit_cap || $post_comment_cap != $attachment->post_comment_cap ) {
 			// use $wpdb instead of wp_update_post to avoid inifinite do_action
 			global $wpdb;
 			$data = array(
@@ -234,33 +240,46 @@ class UndisclosedEditPost {
 			}
 		}
 		
-		if ( ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_view_cap' ) ) && ( $post_type_object->public || $post_type_object->show_ui ) && $post->post_type != 'attachment' ) { 
+		if ( self::can_edit_view_cap( $post->post_type , $post_type_object ) ) { 
 			?><div class="disclosure-view-select misc-pub-section">
-				<label for="post_view_cap-select"><strong><?php _e( 'Who can read:' , 'wpundisclosed') ?></strong></label><br />
+				<label for="post_view_cap-select"><strong><?php _e( 'Who can read:' , 'wp-access-areas') ?></strong></label><br />
 				<?php 
 					self::access_area_dropdown( $rolenames , $groups , $post->post_view_cap , 'post_view_cap' );
 				?>
 			</div><?php
 		}
-		if ( ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_edit_cap' ) ) ) {
+		if ( self::can_edit_edit_cap() ) {
 			?><div class="disclosure-edit-select misc-pub-section">
-				<label for="post_edit_cap-select"><strong><?php _e( 'Who can edit:' , 'wpundisclosed') ?></strong></label><br />
+				<label for="post_edit_cap-select"><strong><?php _e( 'Who can edit:' , 'wp-access-areas') ?></strong></label><br />
 				<?php 
 					self::access_area_dropdown( $edit_rolenames , $groups , $post->post_edit_cap , 'post_edit_cap' );
 				?>
 			</div><?php
 		}
-		if ( ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_comment_cap' ) ) && post_type_supports( $post->post_type , 'comments' ) && wpaa_user_can( $post->post_comment_cap ) ) {
+		if ( self::can_edit_comment_cap( $post->post_type ) && wpaa_user_can( $post->post_comment_cap ) ) {
 			?><div class="disclosure-comment-select misc-pub-section">
-				<label for="post_comment_cap-select"><strong><?php _e( 'Who can comment:' , 'wpundisclosed') ?></strong></label><br />
+				<label for="post_comment_cap-select"><strong><?php _e( 'Who can comment:' , 'wp-access-areas') ?></strong></label><br />
 				<?php 
 					self::access_area_dropdown( $rolenames , $groups , $post->post_comment_cap , 'post_comment_cap' );
 				?>
 			</div><?php
 		}
 	}
+	
+	private static function can_edit_view_cap( $post_type , $post_type_object = null ) {
+		if ( is_null( $post_type_object ) )
+			$post_type_object = get_post_type_object( $post_type );
+		return ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_view_cap' ) ) && ( $post_type_object->public || $post_type_object->show_ui ) && $post_type != 'attachment';
+	}
+	private static function can_edit_edit_cap() {
+		return ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_edit_cap' ) );
+	}
+	private static function can_edit_comment_cap( $post_type ) {
+		 return ( ! get_option( 'wpaa_enable_assign_cap' ) || current_user_can( 'wpaa_set_comment_cap' ) ) && post_type_supports( $post_type , 'comments' );
+	}
+	
 	static function disclosure_box_behavior( ) {
-		$post 				= get_post(get_the_ID());
+		$post 			= get_post(get_the_ID());
 		$post_behavior 	= get_post_meta( $post->ID , '_wpaa_post_behavior' , true );
 		if ( ! $post_behavior )
 			$post_behavior = get_option('wpaa_default_behavior');
@@ -270,7 +289,7 @@ class UndisclosedEditPost {
 			$post_fallback_page = get_option('wpaa_fallback_page');
 
 		?><div class="disclosure-view-select misc-pub-section"><?php
-			?><p class="description"><?php _e('If somebody tries to view a restricted post directly:' , 'wpundisclosed' ); ?></p><?php
+			?><p class="description"><?php _e('If somebody tries to view a restricted post directly:' , 'wp-access-areas' ); ?></p><?php
 		
 		self::behavior_select( $post_behavior );
 		
@@ -278,7 +297,7 @@ class UndisclosedEditPost {
 		
 		?><div class="disclosure-view-select misc-pub-section"><?php
 		?><label for="_wpaa_fallback_page"><?php
-		_e('Fallback Page','wpundisclosed');
+		_e('Fallback Page','wp-access-areas');
 		?></label><?php
 		// only offer non-restricted pages
 		
@@ -300,12 +319,12 @@ class UndisclosedEditPost {
 				?><option value="<?php $first_item_value ?>"><?php echo $first_item_label ?></option><?php
 			}
 		
-			?><option value="exist" <?php selected($selected_cap , 'exist') ?>><?php _e( 'WordPress default' , 'wpundisclosed' ) ?></option><?php
+			?><option value="exist" <?php selected($selected_cap , 'exist') ?>><?php _e( 'WordPress default' , 'wp-access-areas' ) ?></option><?php
 			if ( strpos( $fieldname , 'post_edit_cap' ) === false ) {
-				?><option value="read" <?php selected($selected_cap , 'read') ?>><?php _e( 'Logged in Users' , 'wpundisclosed' ) ?></option><?php
+				?><option value="read" <?php selected($selected_cap , 'read') ?>><?php _e( 'Logged in Users' , 'wp-access-areas' ) ?></option><?php
 			}
 			
-			?><optgroup label="<?php _e( 'WordPress roles' , 'wpundisclosed') ?>">
+			?><optgroup label="<?php _e( 'WordPress roles' , 'wp-access-areas') ?>">
 			<?php foreach ($roles as $role => $rolename) {
 				if ( ! wpaa_user_can_role( $role ) )
 					continue;
@@ -314,12 +333,12 @@ class UndisclosedEditPost {
 			<?php } ?>
 			</optgroup>
 			<?php if ( count($groups) ) { ?>
-				<optgroup label="<?php _e( 'Users with Access to' , 'wpundisclosed') ?>">
+				<optgroup label="<?php _e( 'Users with Access to' , 'wp-access-areas') ?>">
 				<?php foreach ($groups as $group=>$groupname) { 
 					if ( ! wpaa_user_can_accessarea($group) )
 						continue;
 					?>
-					<option value="<?php echo $group ?>" <?php selected($selected_cap , $group) ?>><?php _e( $groupname , 'wpundisclosed' ) ?></option>
+					<option value="<?php echo $group ?>" <?php selected($selected_cap , $group) ?>><?php _e( $groupname , 'wp-access-areas' ) ?></option>
 				<?php } /* foreach( $groups ) */ ?>
 				</optgroup>
 			<?php }  /* if count( $groups ) */ ?>
@@ -353,15 +372,15 @@ class UndisclosedEditPost {
 		$behaviors = array(
 			array( 
 				'value'	=> '404',
-				'label' => __( 'Show 404' , 'wpundisclosed'),
+				'label' => __( 'Show 404' , 'wp-access-areas'),
 			),
 			array(
 				'value'	=> 'page',
-				'label' => __( 'Redirect to the fallback page.' , 'wpundisclosed'),
+				'label' => __( 'Redirect to the fallback page.' , 'wp-access-areas'),
 			),
 			array(
 				'value'	=> 'login',
-				'label' => __( 'If not logged in, redirect to login. Otherwise redirect to the fallback page.' , 'wpundisclosed'),
+				'label' => __( 'If not logged in, redirect to login. Otherwise redirect to the fallback page.' , 'wp-access-areas'),
 			),
 		);
 
@@ -373,6 +392,7 @@ class UndisclosedEditPost {
 			<br /></label><?php
 		}
 	}
+	
 	// --------------------------------------------------
 	// Quick Edit hook callback
 	// --------------------------------------------------
@@ -381,12 +401,14 @@ class UndisclosedEditPost {
 		// enqueue
 		self::_edit_fields( $column_name, $post_type , $post , null );
 	}
+	
 	// --------------------------------------------------
 	// Bulk Edit hook callback
 	// --------------------------------------------------
 	static function bulk_edit_fields( $column_name, $post_type ) {
 		self::_edit_fields( $column_name, $post_type );
 	}
+	
 	// --------------------------------------------------
 	// Quick/Bulk Edit html
 	// --------------------------------------------------
@@ -417,30 +439,32 @@ class UndisclosedEditPost {
 				}
 			}
 			?><fieldset class="inline-edit-col-access-areas inline-edit-col-left">
-				<h3><?php _e('Access','wpundisclosed') ?></h3>
+				<h3><?php _e('Access','wp-access-areas') ?></h3>
 				<div class="inline-edit-col"><?php
-					if ( $post_type_object->public ) {
+					if ( self::can_edit_view_cap( $post_type , $post_type_object ) ) {
 						?><div class="inline-edit-group">
 							<label>
-								<span class="title"><?php _e( 'Read:' , 'wpundisclosed') ?></span>
+								<span class="title"><?php _e( 'Read:' , 'wp-access-areas') ?></span>
 								<?php 
 								self::access_area_dropdown( $rolenames , $groups , $view_cap , 'post_view_cap' , $first_item_value , __( '&mdash; No Change &mdash;' ) );
 								?>
 							</label>
 						</div><?php
 					}
-					?><div class="inline-edit-group">
-						<label>
-							<span class="title"><?php _e( 'Edit:' , 'wpundisclosed') ?></span>
-							<?php 
-							self::access_area_dropdown( $edit_rolenames , $groups , $edit_cap , 'post_edit_cap'  , $first_item_value , __( '&mdash; No Change &mdash;' )  );
-							?>
-						</label>
-					</div><?php
-					if ( post_type_supports( $post_type , 'comments' ) ) {
+					if ( self::can_edit_edit_cap() ) {
 						?><div class="inline-edit-group">
 							<label>
-								<span class="title"><?php _e( 'Comment:' , 'wpundisclosed') ?></span>
+								<span class="title"><?php _e( 'Edit:' , 'wp-access-areas') ?></span>
+								<?php 
+								self::access_area_dropdown( $edit_rolenames , $groups , $edit_cap , 'post_edit_cap'  , $first_item_value , __( '&mdash; No Change &mdash;' )  );
+								?>
+							</label>
+						</div><?php
+					}
+					if ( self::can_edit_comment_cap( $post_type ) ) {
+						?><div class="inline-edit-group">
+							<label>
+								<span class="title"><?php _e( 'Comment:' , 'wp-access-areas') ?></span>
 								<?php 
 								self::access_area_dropdown( $rolenames , $groups , $comment_cap , 'post_comment_cap'  , $first_item_value , __( '&mdash; No Change &mdash;' ) );
 								?>
@@ -467,8 +491,8 @@ class UndisclosedEditPost {
 		foreach ($columns as $k=>$v) {
 			$cols[$k] = $v;
 			if ($k == $after ) {
-				$cols['view_cap'] = __('Visible to','wpundisclosed');
-				$cols['comment_cap'] = __('Commentable to','wpundisclosed');
+				$cols['view_cap'] = __('Visible to','wp-access-areas');
+				$cols['comment_cap'] = __('Commentable to','wp-access-areas');
 			}
 		}
 		return $cols;
@@ -486,7 +510,7 @@ class UndisclosedEditPost {
 		foreach ($columns as $k=>$v) {
 			$cols[$k] = $v;
 			if ( $k == $after ) {
-				$cols['view_cap'] = __('Visible to','wpundisclosed');
+				$cols['view_cap'] = __('Visible to','wp-access-areas');
 			}
 		}
 		return $cols;
@@ -504,7 +528,7 @@ class UndisclosedEditPost {
 		foreach ($columns as $k=>$v) {
 			$cols[$k] = $v;
 			if ( $k == $after ) {
-				$cols['comment_cap'] = __('Commentable to','wpundisclosed');
+				$cols['comment_cap'] = __('Commentable to','wp-access-areas');
 			}
 		}
 		return $cols;
@@ -522,7 +546,7 @@ class UndisclosedEditPost {
 		foreach ($columns as $k=>$v) {
 			$cols[$k] = $v;
 			if ( $k == $after ) {
-				$cols['edit_cap'] = __('Editable for','wpundisclosed');
+				$cols['edit_cap'] = __('Editable for','wp-access-areas');
 			}
 		}
 		return $cols;
@@ -536,19 +560,19 @@ class UndisclosedEditPost {
 // 		var_dump($column,current_filter());
 		switch ( $column ) {
 			case 'view_cap':
-				$names = array_merge(array('exist' => __( 'Everybody' , 'wpundisclosed' ), 'read' => __( 'Blog users' , 'wpundisclosed' )) , UndisclosedUserlabel::get_label_array( ), $wp_roles->get_names());
+				$names = array_merge(array('exist' => __( 'Everybody' , 'wp-access-areas' ), 'read' => __( 'Blog users' , 'wp-access-areas' )) , UndisclosedUserlabel::get_label_array( ), $wp_roles->get_names());
 				$names[''] = $names['exist'];
 				$val = get_post($post_ID)->post_view_cap;
 				_e($names[$val]);
 				break;
 			case 'comment_cap':
-				$names = array_merge(array('exist' => __( 'Everybody' , 'wpundisclosed' ), 'read' => __( 'Blog users' , 'wpundisclosed' )) , UndisclosedUserlabel::get_label_array( ), $wp_roles->get_names());
+				$names = array_merge(array('exist' => __( 'Everybody' , 'wp-access-areas' ), 'read' => __( 'Blog users' , 'wp-access-areas' )) , UndisclosedUserlabel::get_label_array( ), $wp_roles->get_names());
 				$names[''] = $names['exist'];
 				$val = get_post($post_ID)->post_comment_cap;
 				_e($names[$val]);
 				break;
 			case 'edit_cap':
-				$names = array_merge(array('exist' => __( 'Everybody' , 'wpundisclosed' ), 'read' => __( 'Blog users' , 'wpundisclosed' )) , UndisclosedUserlabel::get_label_array( ), $wp_roles->get_names());
+				$names = array_merge(array('exist' => __( 'Everybody' , 'wp-access-areas' ), 'read' => __( 'Blog users' , 'wp-access-areas' )) , UndisclosedUserlabel::get_label_array( ), $wp_roles->get_names());
 				$names[''] = $names['exist'];
 				$val = get_post($post_ID)->post_edit_cap;
 				_e($names[$val]);
